pyproject.toml
[tool.poetry]
name = "oc-student-sentiment"
version = "0.1.0"
description = ""
authors = ["sidux <ahmed.lebbada@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
transformers = "^4.48.0"
tensorflow = "^2.18.0"
tf-keras = "^2.18.0"
jinja2 = "^3.1.5"
opencensus-ext-azure = "^1.1.14"
opencensus-ext-fastapi = "^0.1.0"
opencensus-ext-logging = "^0.1.1"
python-dotenv = "^1.0.1"
middleware = "^1.2.3"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
azure-monitor-opentelemetry = "^1.6.4"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
